+++ Models/Json/DmartJsonContext.cs
@@ -41,6 +41,7 @@ namespace Dmart.Models.Json;
 [JsonSerializable(typeof(Status))]
 [JsonSerializable(typeof(UserLoginRequest))]
 [JsonSerializable(typeof(SendOTPRequest))]
+[JsonSerializable(typeof(VerifyContactRequest))]
 [JsonSerializable(typeof(PasswordResetConfirm))]
 [JsonSerializable(typeof(UserCreateBody))]
 [JsonSerializable(typeof(RegisterRequest))]